libfdt: overlay: Add export-symbols support
The export symbols node adds some additional symbols that can be used
in the symbols resolution. The resolver tries to match unresolved
symbols first using the export symbols node and, if a match is not
found, it tries the normal route and searches the __symbols__ node.

Contrary to symbols available in the global __symbols__ node, symbols
listed in the export symbols node can be considered as local symbols.
Indeed, they can be changed depending on the node the overlay is going
to be applied to and are only visible from the current node properties.

export-symbols are phandle based as opposed to global __symbols__. This
allows for much simpler use with overlays as opposed to __symbols__
where paths require resizing of overlays as show in [0].

The phandle resolution in overlay is now broken into 2 steps:
- fragment target phandles
- All other phandles

The reason this is required is because to find the export-symbols node
in the target node, we first need to resolve the target phandles. Here
is an example of overlay that will generate the target __fixup__ after
the property fixup, and thus fail without this 2 step resolution.

```
/dts-v1/;
/plugin/;

&connectora {
  prop = <&gpio_connector>;
};

&connectorb {
  prop = <&gpio_connector>;
};
```

Using fdtdump, we get the following:

```
**** fdtdump is a low-level debugging tool, not meant for general use.
**** If you want to decompile a dtb, you probably want
****     dtc -I dtb -O dts <filename>

/dts-v1/;
// magic:               0xd00dfeed
// totalsize:           0x1b1 (433)
// off_dt_struct:       0x38
// off_dt_strings:      0x180
// off_mem_rsvmap:      0x28
// version:             17
// last_comp_version:   16
// boot_cpuid_phys:     0x0
// size_dt_strings:     0x31
// size_dt_struct:      0x148

/ {
    fragment@0 {
        target = <0xffffffff>;
        __overlay__ {
            prop = <0xffffffff>;
        };
    };
    fragment@1 {
        target = <0xffffffff>;
        __overlay__ {
            prop = <0xffffffff>;
        };
    };
    __fixups__ {
        connectora = "/fragment@0:target:0";
        gpio_connector = "/fragment@0/__overlay__:prop:0", "/fragment@1/__overlay__:prop:0";
        connectorb = "/fragment@1:target:0";
    };
};
```

[0]: https://lore.kernel.org/devicetree-compiler/6b2dba90-3c52-4933-88f3-b47f96dc7710@beagleboard.org/T/#m8259c8754f680b9da7b91f7b7dd89f10da91d8ed

Signed-off-by: Ayush Singh <ayush@beagleboard.org>